DefaultFormsController refactoring
