+++ web-client/src/main/java/lsfusion/gwt/client/form/view/ModalForm.java
@@ -1,19 +1,15 @@
 package lsfusion.gwt.client.form.view;
 
-import com.google.gwt.core.client.Scheduler;
-import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.ui.Widget;
-import lsfusion.gwt.client.base.Dimension;
 import lsfusion.gwt.client.base.view.ResizableModalWindow;
-import lsfusion.gwt.client.form.controller.DefaultFormsController;
-import lsfusion.gwt.client.form.controller.GFormController;
+import lsfusion.gwt.client.form.controller.FormsController;
 import lsfusion.gwt.client.view.MainFrame;
 
 import static java.lang.Math.min;
 
 public class ModalForm extends FormContainer<ResizableModalWindow> {
 
-    public ModalForm(DefaultFormsController formsController) {
+    public ModalForm(FormsController formsController) {
         super(formsController);
     }
